% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drilling_data.R
\docType{package}
\name{DrillingData}
\alias{DrillingData}
\alias{DrillingData-package}
\title{A package of drilling data sets}
\description{
\describe{
  \item{FYH_drill}{Diamond drill core data from Western Australia}
}
}
